requirements.txt
Django==3.1.7
django-allauth==0.44.0
django-cors-headers==3.7.0
django-debug-toolbar==3.2
django-filter==2.4.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
Pillow==8.1.2
psycopg2-binary==2.8.6
django-environ=0.4.5
